LPS-56065-PortalSettingsPortletToOSGI - Move EditLDAPServerAction.java, no logic changes
